All of Semblance's bot listed sites are added
