+++ index.js
@@ -12,9 +12,8 @@ const Semblance = require('./structures/Semblance'),
 	}),
 	// Database connection import
 	{ connect } = require('mongoose'),
-	// Event handlers
+	// Client event handlers
 	checkTweet = require('./events/checkTweet'),
-	discordBoat = require('./events/discordBoat'),
 	interactionCreate = require('./events/interactionCreate'),
 	message = require('./events/message'),
 	messageDelete = require('./events/messageDelete'),
@@ -22,7 +21,13 @@ const Semblance = require('./structures/Semblance'),
 	messageReactionRemove = require('./events/messageReactionRemove'),
 	messageUpdate = require('./events/messageUpdate'),
 	ready = require('./events/ready'),
-	topGG = require('./events/topGG'),
+	// Bot listing event handlers
+	botListSpace = require('./events/botListingEvents/botListSpace'),
+	botsForDiscord = require('./events/botListingEvents/botsForDiscord'),
+	discordBoat = require('./events/botListingEvents/discordBoat'),
+	discordBotList = require('./events/botListingEvents/discordBotList'),
+	discordBotsGG = require('./events/botListingEvents/discordBotsGG'),
+	topGG = require('./events/botListingEvents/topGG'),
 	// Ping web host (Heroku)
 	stayActive = require('./stayActive.js');
 
@@ -34,6 +39,13 @@ messageReactionAdd(client);
 messageReactionRemove(client);
 messageUpdate(client);
 ready(client);
+// Listen to bot listing events
+botListSpace(client);
+botsForDiscord(client);
+discordBoat(client);
+discordBotList(client);
+discordBotsGG(client);
+topGG(client);
 // Check for Tweet from ComputerLunch
 setInterval(() => checkTweet(client), 2000);
