+++ package.json
@@ -2,22 +2,23 @@
   "name": "semblance",
   "version": "1.0.0",
   "description": "The AI that runs the Reality Engine",
-  "main": "bot.js",
+  "main": "index.js",
   "author": "SirH",
   "dependencies": {
     "boats.js": "^2.5.1",
+    "botlist.space": "^3.3.1",
     "cron": "^1.8.2",
     "dblapi.js": "^2.4.1",
     "discord.js": "^12.5.1",
     "ffmpeg": "0.0.4",
     "ffmpeg-static": "^4.2.7",
-    "mongoose": "^5.11.18",
+    "mongoose": "^5.11.19",
     "node-fetch": "^2.6.1",
     "node-gyp": "^7.1.2",
     "opusscript": "0.0.7",
-    "pm2": "^4.5.4",
+    "pm2": "^4.5.5",
     "request": "^2.88.2",
-    "tmi.js": "^1.7.1",
+    "tmi.js": "^1.7.5",
     "twitter": "^1.7.1",
     "winston": "^3.3.3",
     "ytdl-core-discord": "^1.2.5"
